+++ src/clj/clj_toolbox/base64.clj
@@ -1,0 +1,8 @@
+(ns clj-toolbox.base64
+  (:import
+    [java.util Base64]))
+
+(defn encode-str
+  "base64 encodes the input string. Returns a string"
+  [^String s]
+  (.encodeToString (Base64/getEncoder) (.getBytes s)))